Store install dir in settings struct
Remove duplicate "central" field tracking the install dir, which was
redundant as the install dir was already stored in a publicly accesible
field of the settings struct. Also pass a message to the settings
struct when the main gui receives the CloseSettings message to trigger
directory validation in settings, instead of in gui.