Add ModDescription for displaying descriptions
Provides a struct with the same construction as other iced components.
This component initially begins without a connected mod entry, and is
updated with a new entry whenever an entry in the mod_list is clicked.
The description box thus automatically updates to display the
description of whatever mod was most recently selected.