Make compress-tools platform specific
This allows the creation of statically linked executables, and removes
the deployment nightmare that would be trying to include libarchive.
This does mean the loss of 7z support on Windows. This is accomplished
by using conditional compilation to make the 7z decompression method
compile down to a null-op when compiling for windows.