Handle nested mod folders, existing directories
Handles moving and renaming mod folders when they are nested within
other folders in a given archive. Also now handles occasions where the
a directory with the same name as the directory a mod will be extracted
to already exists - in this scenario, a popup will ask the user if they
wish to cancel the archive extraction, or replace the existing
directory (ie: delete it entirely).