Allow installing mods from an already decompressed folder Fixes #4
This feature enables the option to install mods from a folder that
already exists.

The features works by asking a user for a directory containing the mod,
which can be a parent directory to the mod directory. The app then
copies each file in the mod directory to a new directory in the mods
directory and recreates each directory in the given mod directory as
well. The given/original mod directory is then recursively deleted.
The reason a copy-delete approach is taken is because: 1) fs::rename
does not work across mount points, and 2) there is no std recursive
directory copy available in std::fs or elsewhere.