Make archive_handler error sendable
Makes archive_handler errors sendable at the cost of removing Unrarr
error information. Unrarr errors are now "opaque", in that they are
still noted and raised, but the exact details are discarded as they
reference unsendable objects.

If this becomes an issue in the future it may be possible to wrap
Unrarr errors more completely so that error details can be converted to
a sendable form and passed on.