Add async archive install methods
The install module is a new module that will provide async mod install
methods, generally for handling I/O intensive install operations. This
will eventually include folder based installs, and at the moment covers
single archive installs.

The intention is for the async single archive handler to be invoked
as many times as needed when installing multiple mods, with tokio
sorting things out for us. This means that there is now a way to not
only parallelize the entire job of archive installs, but every archive
install job in the entire batch (hopefully tokio is smart enough to
only parallelize this as much as is actually possible given the cores
of the executing machine).

The intended pattern is now:
 - (Synchronously) receive list of mod archives to install from
 - (Synchronously) create install jobs pointing to each archive dir
 - (Asynchronously) crunch all install jobs
 - (Synchronously) respond to each return value returned from the
install jobs.