Add console message for rare mod list error
Generally, we don't expect the mod folder parsing method to fail (most
of the ifs and lets are to cover optionals that are guaranteed to have
contents elsewhere). This doesn't stop bad permissions and stuff, but
those should crop up elsewhere anyway.

What _is_ an issue is if certain files aren't there. In this case, the
way the method is written assumes the enabled_mods.json file is always
present. There's two invocations that depend on the existence of the
file and its contents. The method previously failed silently if this
was not the case (much to my annoyance).