Switch to async mod install from archive
Installing mods from archives is now performed asynchronously, using
the implementation in the install module.

The list of mods archives to install from is collected, then converted
to a list of Commands performing the async mod install method pointed
to the given paths.

A message is presented whenever a result from an install is received.
This includes informing the user that the installation has succeeded,
or that there has been an error.

A unique case exists where the error is due to the directory a mod will
be installed to already existing. This error will prompt a query to the
user asking if they would like to overwrite/delete the existing
directory. Any further failures are treated as any other error.