Implement install from folder as async
Reworked the install_archive function to also install folders. Both
do very similar work with path manipulation, so there was little reason
not to merge the functionality into the existing function than to
duplicate it in another method - though this might change in the future
if it becomes unhelpful. The folder installing behaviour is switched by
a boolean parameter.

Also adds the installed mod's name to the error type DeleteError, as
it indicates a partial success, where the mod is installed/copied but
the temp/original folder could not be deleted.