Now display retrieved update information
Mod version cells have different backgrounds depending on what update
data on the mod has been retrieved. Amber indicates an update is
available. Blue indicates mod is up to date (according to remote). Red
indicates there was an error retrieving update data. No colour
indicates the mod does not include a version/does not support version
checking.

Tooltips present additional information, describing how the mod is out
of date/if it is up to date/that there was an error. The local and
remote versions are listed if there is a difference.