Fix incorrect non-numeric handling in settings
When modifying a vmparams value, it was intended that typing in a non-
numeric value (like a char) would make no change, and that removing the
input entirely/making it empty would set it to a default value.

The former was incorrectly implemented, and caused any typing of non-
digits to cause the field to be set to zero. This is now fixed.