Fix handling of existing ID collision/upgrading
Handing the temporary folder off to the standard install code was not
correct, as the install function has a number of embedded assumptions
(ie: the given folder is reasonably named, or has a nested child folder
that actually contains the mod. It could probably do with changed
anyway...).

Fixed by using custom handling of ID collisions in the match branch of
the mod_list handler, which just runs a rename and deletes any empty
parent directories that were left behind.