Fix more mod upgrade/replace bugs
When encountering existing folder, delete mod from internal mods list
if we knew it was installed to the deleted path - this prevents the ID
collision handler (probably) always running after the folder collision
handler.
Fix the ID collision handler to continue if the existing mod is no
longer installed (ie: it's directory was deleted in the interm) and to
call parse_mod_folder on success, to ensure internal mod list matches
the state of the actual mod folder - this prevents cases where a
successive upgrades could result in the recorded installation path of a
mod would no longer reflect reality, and allow for multiple versions of
a mod to be installed in parallel without detection.

Improved error messaging for ID collisions so the next time something
goes wrong here it's easier for users to report and to start debugging.
Also now generate a success popup like with other success scenarios, so
that users know the upgrade/replace by ID operation has succeeded.