Prevent resizing of Enabled column
This column is already intentionally shrunk, and I can't see a reason
someone would want to expand it (and by doing this I'm not giving anyone
the chance to find one I need to support).
This also circumvents a janky bug where it wasn't possible to shrink
the column back down to its original size after expanding it - I
believe there's some kind of minimum sizing shenanigans going on there.