Make config live in platform specific config storage
Adds directories as a dependency.

Make gui::config::path async and return either a path representing the
current platform's app config location (ie: Application Support or
AppData, etc) joined with `config.json`. The path function will also
attempt to create the config directory if it does not exist if
instructed to - currently this only occurs when saving the config file.