Add tempfile as a dependency
This crate utilises OS provided "temporary" files and directories
which, if not intentionally persisted, are automatically removed as
soon as the handler for the file/folder moves out of scope.

This allows us to write out files "temporarily", examine their
contents, and check if the new mod would either conflict by ID or
folder name with an existing mod.