Remove find_mountpoint
Whilst it does appear to be functioning correctly (finding mount
points) it appears that some things have changed since the crate was
developed (unsurprisingly) rendering it potentially unreliable.

Specifically, the functions assume that all paths given to it will
contain their mout points, however it appears that, at least in the
case of macOS, the drive/mount point of the path is disguised as the
root of the entire filesystem, causing a mismatch.