Add script to rewrite macos binary deps
This script rewrites the dylib dependency paths of the binary to point
to local copies, with search paths relative to the binary itself, and
copies the dylibs to the expected location. Also adds a build step for
macos to run the script.

This should hopefully make macos releases truly "portable", thus
solving the crash users have when they don't install libarchive through
brew.