Add mod auto-update functionality Fixes #102
Adds a new column indicating if a mod supports automatic updates -
where the update is downloaded and applied automatically.
Clicking the button in the new column opens a new modal dialog asking
if the user is sure they want to continue.
Amends util:get_master_version to always return a ModVersion on
success, and moved determination of whether a mod is up to date to
ModList::update.