Implement download progress bars.
Also provides a new "hover" controller implementation - this
implementation should be generic enough to be applied as a controller to
any widget and provide a "pointer on hover" effect.

All downloads now generate a progress bar. Progress update events are
load balanced through a dedicated thread which handles batching updates
to the main UI thread - the download threads send progress updates over
an mpsc channel to the balancing thread.